chore:  Refactor Graph2 to be more generic
Not 100% complete, but GraphT has reduced dependency on IVertex/Vertex
