+++ packages/graph/package.json
@@ -60,13 +60,14 @@
     },
     "devDependencies": {
         "@hpcc-js/bundle": "^2.11.3",
-        "@hpcc-js/wasm": "1.15.7",
+        "@hpcc-js/wasm": "1.16.5",
         "@types/d3-transition": "1.3.2",
         "@types/dagre": "0.7.48",
         "d3-array": "^1",
         "d3-drag": "^1",
         "d3-force": "^1",
         "d3-geo": "^1",
+        "d3-interpolate-path": "2.3.0",
         "d3-sankey": "^0",
         "d3-scale": "^1",
         "d3-selection": "^1",
