+++ packages/graph/src/graph2/index.ts
@@ -1,3 +1,4 @@
+export * from "./graphT";
 export * from "./graph";
 export * from "./dataGraph";
 export * from "./sankeyGraph";
